Move reference tests to a separate class
